refactor(content): extract parsing and series metadata, breaking the parse↔series cycle
parseMarkdownFile/parseRstFile inherit series titles and authors, and the
series query functions need the parsers — co-location made that look like
a cycle. It never was one: getSeriesTitle/getSeriesAuthors read series
index files directly via gray-matter/parseRstDocument. content/series-metadata.ts
now owns index discovery (mixed-format and duplicate-slug throws intact)
plus title/author reads, and content/parse.ts owns PostSchema, both file
parsers, the rST toctree helpers, and the Python-renderer availability
singleton together with its test hooks. The chain is acyclic
(series-metadata → parse → queries) and a dependency-graph guard test
keeps it that way — ESM cycles fail silently as undefined bindings, so
lint+test must catch them. getAllPosts' inline batch-render block becomes
parseRstPostEntries(), keeping the singleton private to parse.ts.
