+++ src/lib/markdown.ts
@@ -12,7 +12,7 @@ import {
   generateExcerpt,
   getHeadings,
 } from './text-metrics';
-import type { PostData, CollectionItem, CollectionContext, Heading } from './content/types';
+import type { PostData, CollectionContext, Heading } from './content/types';
 import {
   contentDirectory,
   pagesDirectory,
@@ -21,92 +21,20 @@ import {
   flowsDirectory,
   notesDirectory,
   readUtf8File,
-  isMarkdownFilename,
-  isRstFilename,
   parseSlugAndDate,
-  assertSafeSeriesSlug,
 } from './content/io';
 import { getCacheEnvKey } from './content/cache';
-import { parseRstDocument, RstParseError } from './rst';
-import { renderRstFile, renderRstFilesBatch, type RenderedRstDocument } from './rst-renderer';
-
-const ExternalLinkSchema = z.object({
-  name: z.string(),
-  url: z.string().url(),
-});
-
-const CollectionItemSchema = z.union([
-  z.object({
-    series: z.string(),
-    exclude: z.array(z.string()).optional(),
-    label: z.string().optional(),
-  }).strict(),
-  z.object({
-    post: z.string(),
-    label: z.string().optional(),
-  }).strict(),
-]);
-
-const PostSchema = z.object({
-  title: z.string(),
-  date: z.union([z.string(), z.date()]).transform(val => new Date(val).toISOString().split('T')[0]).optional(),
-  subtitle: z.string().optional(),
-  excerpt: z.string().optional(),
-  category: z.string().optional().default('Uncategorized'),
-  tags: z.array(z.string()).optional().default([]),
-  authors: z.array(z.string()).optional(),
-  author: z.string().optional(),
-  layout: z.string().optional().default('post'),
-  series: z.string().optional(),
-  coverImage: z.string().optional(),
-  sort: z.enum(['date-desc', 'date-asc', 'manual']).optional().default('date-desc'),
-  posts: z.array(z.string()).optional(),
-  type: z.literal('collection').optional(),
-  items: z.array(CollectionItemSchema).optional(),
-  featured: z.boolean().optional().default(false),
-  pinned: z.boolean().optional().default(false),
-  draft: z.boolean().optional().default(false),
-  latex: z.boolean().optional().default(false),
-  toc: z.boolean().optional().default(true),
-  commentable: z.boolean().optional(),
-  externalLinks: z.array(ExternalLinkSchema).optional().default([]),
-  redirectFrom: z.array(z.string()).optional().default([]),
-}).superRefine((data, ctx) => {
-  if (data.type === 'collection' && (!data.items || data.items.length === 0)) {
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      path: ['items'],
-      message: 'Collections require at least one item.',
-    });
-  }
-  if (data.type !== 'collection' && data.items) {
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      path: ['items'],
-      message: '`items` is only valid when `type` is "collection".',
-    });
-  }
-});
-
-type SeriesFormat = 'markdown' | 'rst';
-
-interface SeriesIndexInfo {
-  format: SeriesFormat;
-  fullPath: string;
-}
-
-interface SeriesContentEntry {
-  fullPath: string;
-  slug: string;
-  dateFromFileName?: string;
-}
-
-interface PendingRstPostEntry {
-  fullPath: string;
-  slug: string;
-  dateFromFileName?: string;
-  seriesSlug?: string;
-}
+import {
+  resolveSeriesIndexInfo,
+  getSeriesContentEntries,
+  getSeriesAuthors,
+} from './content/series-metadata';
+import {
+  parseMarkdownFile,
+  parseRstFile,
+  parseRstPostEntries,
+  type RstPostEntry,
+} from './content/parse';
 
 const postsCache = new Map<string, PostData[]>();
 const pagesCache = new Map<string, PostData[]>();
@@ -122,145 +50,6 @@ const featuredSeriesCache = new Map<string, Record<string, PostData[]>>();
 const seriesLatestDateCache = new Map<string, Map<string, string>>();
 const collectionPostsCache = new Map<string, Map<string, PostData[]>>();
 const collectionsForPostCache = new Map<string, Map<string, CollectionContext[]>>();
-const seriesAuthorsCache = new Map<string, Map<string, string[] | null>>();
-const seriesTitleCache = new Map<string, Map<string, string | undefined>>();
-let pythonRstRendererAvailable: boolean | null = null;
-
-const PYTHON_RUNTIME_UNAVAILABLE_PATTERN = /docutils|No module named|python(?:3)? .*not found|interpreter not found|ENOENT.*python/i;
-
-function isPythonRuntimeUnavailable(error: unknown): boolean {
-  if (!(error instanceof Error)) return false;
-  if (error.message.includes('__RST_FALLBACK__')) return true;
-  if (error.message.includes('rST file not found')) return false;
-  return PYTHON_RUNTIME_UNAVAILABLE_PATTERN.test(error.message);
-}
-
-function getRstImageBaseSlug(fullPath: string, slug: string): string {
-  const isRootFlatPost = path.basename(fullPath) !== 'index.rst' &&
-    path.dirname(fullPath) === contentDirectory;
-  return isRootFlatPost ? 'posts' : `posts/${slug}`;
-}
-
-function isSeriesIndexRst(fullPath: string, slug: string, seriesName?: string): boolean {
-  return Boolean(
-    seriesName &&
-    slug === seriesName &&
-    (path.basename(fullPath) === 'index.rst' || path.basename(fullPath) === 'README.rst')
-  );
-}
-
-function slugFromRstToctreeTarget(target: string): string | null {
-  const trimmed = target.trim();
-  if (!trimmed || trimmed.startsWith(':')) return null;
-  if (/^[a-z]+:\/\//i.test(trimmed) || trimmed.startsWith('/')) return null;
-
-  const withoutAnchor = trimmed.split('#')[0]?.split('?')[0]?.trim();
-  if (!withoutAnchor) return null;
-
-  const normalized = withoutAnchor.replace(/\\/g, '/').replace(/^\.\//, '').replace(/\/+$/, '');
-  if (!normalized || normalized.startsWith('../')) return null;
-
-  const withoutExt = normalized.replace(/\.rst$/i, '');
-  const parts = withoutExt.split('/').filter(Boolean);
-  if (parts.length === 0) return null;
-
-  const last = parts[parts.length - 1];
-  if (last === 'index' || last === 'README') {
-    return parts.length > 1 ? parts[parts.length - 2] : null;
-  }
-
-  return last;
-}
-
-function extractRstToctreePosts(source: string): string[] {
-  const lines = source.replace(/\r\n?/g, '\n').split('\n');
-  const posts: string[] = [];
-  const seen = new Set<string>();
-
-  for (let i = 0; i < lines.length; i++) {
-    if (!/^\s*\.\.\s+toctree::\s*$/.test(lines[i])) continue;
-
-    i++;
-    while (i < lines.length) {
-      const line = lines[i];
-      if (!line.trim()) {
-        i++;
-        continue;
-      }
-      if (!/^\s+/.test(line)) {
-        i--;
-        break;
-      }
-
-      const trimmed = line.trim();
-      if (!trimmed.startsWith(':')) {
-        const slug = slugFromRstToctreeTarget(trimmed);
-        if (slug && !seen.has(slug)) {
-          seen.add(slug);
-          posts.push(slug);
-        }
-      }
-      i++;
-    }
-  }
-
-  return posts;
-}
-
-function shouldUsePythonRstRenderer(): boolean {
-  if (process.env.AMYTIS_ENABLE_PYTHON_RST === '1') return true;
-  if (process.env.AMYTIS_ENABLE_PYTHON_RST === '0') return false;
-  return process.env.NODE_ENV !== 'test';
-}
-
-/**
- * Read explicitly configured authors from a series index file's frontmatter.
- * Returns null if no authors are configured (as opposed to the default fallback).
- */
-export function getSeriesAuthors(seriesSlug: string): string[] | null {
-  const cacheKey = getCacheEnvKey();
-  let bySlug = seriesAuthorsCache.get(cacheKey);
-  if (!bySlug) {
-    bySlug = new Map();
-    seriesAuthorsCache.set(cacheKey, bySlug);
-  }
-  if (bySlug.has(seriesSlug)) return bySlug.get(seriesSlug) ?? null;
-
-  const indexInfo = resolveSeriesIndexInfo(seriesSlug);
-  if (!indexInfo) {
-    bySlug.set(seriesSlug, null);
-    return null;
-  }
-
-  if (indexInfo.format === 'rst') {
-    const parsed = parseRstDocument(readUtf8File(indexInfo.fullPath));
-    if (parsed.metadata.authors && parsed.metadata.authors.length > 0) {
-      bySlug.set(seriesSlug, parsed.metadata.authors);
-      return parsed.metadata.authors;
-    }
-    if (parsed.metadata.author && typeof parsed.metadata.author === 'string') {
-      const authors = [parsed.metadata.author];
-      bySlug.set(seriesSlug, authors);
-      return authors;
-    }
-    bySlug.set(seriesSlug, null);
-    return null;
-  }
-
-  const { data } = matter(readUtf8File(indexInfo.fullPath));
-  if (data.authors && Array.isArray(data.authors) && data.authors.length > 0) {
-    const authors = data.authors as string[];
-    bySlug.set(seriesSlug, authors);
-    return authors;
-  }
-  if (data.author && typeof data.author === 'string') {
-    const authors = [data.author as string];
-    bySlug.set(seriesSlug, authors);
-    return authors;
-  }
-  bySlug.set(seriesSlug, null);
-  return null;
-}
 
 /**
  * Resolve display authors for a series: explicit series authors first,
@@ -281,420 +70,13 @@ export function resolveSeriesAuthors(slug: string, posts: PostData[]): string[] 
     .map(([name]) => name);
 }
 
-function resolveUniqueSeriesIndex(seriesSlug: string, format: SeriesFormat): string | null {
-  assertSafeSeriesSlug(seriesSlug);
-  const seriesPath = path.join(seriesDirectory, seriesSlug);
-  const candidates = format === 'rst'
-    ? ['index.rst', 'README.rst']
-    : ['index.mdx', 'index.md', 'README.mdx', 'README.md'];
-
-  const matches = candidates
-    .map(name => path.join(seriesPath, name))
-    .filter(fullPath => fs.existsSync(fullPath));
-
-  if (matches.length > 1) {
-    throw new Error(
-      `[amytis] Series "${seriesSlug}" has multiple ${format} index files: ${matches.map(match => path.basename(match)).join(', ')}.`
-    );
-  }
-
-  return matches[0] ?? null;
-}
-
-function resolveSeriesIndexInfo(slug: string): SeriesIndexInfo | null {
-  assertSafeSeriesSlug(slug);
-  if (!fs.existsSync(seriesDirectory)) return null;
-  const seriesPath = path.join(seriesDirectory, slug);
-  if (!fs.existsSync(seriesPath) || !fs.statSync(seriesPath).isDirectory()) return null;
-
-  const rstIndex = resolveUniqueSeriesIndex(slug, 'rst');
-  const markdownIndex = resolveUniqueSeriesIndex(slug, 'markdown');
-
-  if (rstIndex && markdownIndex) {
-    throw new Error(
-      `[amytis] Series "${slug}" cannot contain both rST and Markdown index files (${path.basename(rstIndex)} and ${path.basename(markdownIndex)}).`
-    );
-  }
-  if (rstIndex) return { format: 'rst', fullPath: rstIndex };
-  if (markdownIndex) return { format: 'markdown', fullPath: markdownIndex };
-  return null;
-}
-
-function getSeriesContentEntries(seriesSlug: string): SeriesContentEntry[] {
-  const indexInfo = resolveSeriesIndexInfo(seriesSlug);
-  if (!indexInfo) return [];
-
-  const seriesPath = path.join(seriesDirectory, seriesSlug);
-  const seriesItems = fs.readdirSync(seriesPath, { withFileTypes: true });
-  const entries: SeriesContentEntry[] = [];
-  const seenSlugs = new Map<string, string>();
-  const seriesIndexBasenames = new Set(['index.rst', 'README.rst', 'index.md', 'index.mdx', 'README.md', 'README.mdx']);
-
-  for (const item of seriesItems) {
-    if (seriesIndexBasenames.has(item.name)) continue;
-
-    if (item.isFile()) {
-      const isMarkdown = isMarkdownFilename(item.name);
-      const isRst = isRstFilename(item.name);
-      if (!isMarkdown && !isRst) continue;
-
-      const itemFormat: SeriesFormat = isRst ? 'rst' : 'markdown';
-      if (itemFormat !== indexInfo.format) {
-        throw new Error(`[amytis] Series "${seriesSlug}" mixes ${indexInfo.format} and ${itemFormat} files. Offending file: ${item.name}`);
-      }
-
-      const rawName = item.name.replace(/\.(mdx?|rst)$/, '');
-      const { slug, dateFromFileName } = parseSlugAndDate(rawName);
-      const prior = seenSlugs.get(slug);
-      if (prior) {
-        throw new Error(`[amytis] Series "${seriesSlug}" contains duplicate post slug "${slug}" from "${prior}" and "${item.name}".`);
-      }
-      seenSlugs.set(slug, item.name);
-      entries.push({ fullPath: path.join(seriesPath, item.name), slug, dateFromFileName });
-      continue;
-    }
-
-    if (item.isDirectory()) {
-      const folderPath = path.join(seriesPath, item.name);
-      const folderIndexRst = path.join(folderPath, 'index.rst');
-      const folderIndexMdx = path.join(folderPath, 'index.mdx');
-      const folderIndexMd = path.join(folderPath, 'index.md');
-      const hasRst = fs.existsSync(folderIndexRst);
-      const hasMdx = fs.existsSync(folderIndexMdx);
-      const hasMd = fs.existsSync(folderIndexMd);
-      const markdownCount = Number(hasMdx) + Number(hasMd);
-      const totalIndexCount = Number(hasRst) + markdownCount;
-
-      if (totalIndexCount === 0) continue;
-      if (hasRst && markdownCount > 0) {
-        throw new Error(`[amytis] Series "${seriesSlug}" post folder "${item.name}" cannot contain both index.rst and Markdown index files.`);
-      }
-      if (markdownCount > 1) {
-        throw new Error(`[amytis] Series "${seriesSlug}" post folder "${item.name}" cannot contain both index.md and index.mdx.`);
-      }
-
-      const itemFormat: SeriesFormat = hasRst ? 'rst' : 'markdown';
-      if (itemFormat !== indexInfo.format) {
-        throw new Error(`[amytis] Series "${seriesSlug}" mixes ${indexInfo.format} and ${itemFormat} files. Offending folder: ${item.name}`);
-      }
-
-      const { slug, dateFromFileName } = parseSlugAndDate(item.name);
-      const prior = seenSlugs.get(slug);
-      if (prior) {
-        throw new Error(`[amytis] Series "${seriesSlug}" contains duplicate post slug "${slug}" from "${prior}" and "${item.name}".`);
-      }
-      seenSlugs.set(slug, item.name);
-      entries.push({
-        fullPath: hasRst ? folderIndexRst : (hasMdx ? folderIndexMdx : folderIndexMd),
-        slug,
-        dateFromFileName,
-      });
-    }
-  }
-
-  return entries;
-}
-
-function getSeriesTitle(slug: string): string | undefined {
-  const cacheKey = getCacheEnvKey();
-  let bySlug = seriesTitleCache.get(cacheKey);
-  if (!bySlug) {
-    bySlug = new Map();
-    seriesTitleCache.set(cacheKey, bySlug);
-  }
-  if (bySlug.has(slug)) return bySlug.get(slug);
-
-  const indexInfo = resolveSeriesIndexInfo(slug);
-  if (!indexInfo) {
-    bySlug.set(slug, undefined);
-    return undefined;
-  }
-
-  if (indexInfo.format === 'rst') {
-    const parsed = parseRstDocument(readUtf8File(indexInfo.fullPath));
-    if (parsed.metadata.draft === true) {
-      bySlug.set(slug, undefined);
-      return undefined;
-    }
-    bySlug.set(slug, parsed.title);
-    return parsed.title;
-  }
-
-  const { data } = matter(readUtf8File(indexInfo.fullPath));
-  if (data.draft === true) {
-    bySlug.set(slug, undefined);
-    return undefined;
-  }
-  const title = typeof data.title === 'string' ? data.title : undefined;
-  bySlug.set(slug, title);
-  return title;
-}
-
-function parseMarkdownFile(fullPath: string, slug: string, dateFromFileName?: string, seriesName?: string): PostData {
-  const fileContents = readUtf8File(fullPath);
-  const { data: rawData, content } = matter(fileContents);
-  // Flat files directly in content/posts/ share the posts root public directory for images.
-  // Folder-based posts and series posts each have their own public subdirectory.
-  const isRootFlatPost = path.basename(fullPath) !== 'index.mdx' &&
-    path.basename(fullPath) !== 'index.md' &&
-    path.dirname(fullPath) === contentDirectory;
-  const imageBaseSlug = isRootFlatPost ? 'posts' : `posts/${slug}`;
-
-  const parsed = PostSchema.safeParse(rawData);
-  if (!parsed.success) {
-    console.error(`Invalid frontmatter in ${fullPath}:`, parsed.error.format());
-    throw new Error(`Invalid frontmatter in ${fullPath}`);
-  }
-  const data = parsed.data;
-
-  const contentWithoutH1 = content.replace(/^\s*#\s+[^\n]+/, '').trim();
-
-  const effectiveSeriesSlug = data.series || seriesName;
-  let authors: string[] = [];
-  if (data.authors && Array.isArray(data.authors)) {
-    authors = data.authors;
-  } else if (data.author) {
-    authors = [data.author];
-  } else {
-    // Inherit from series if this post belongs to one
-    if (effectiveSeriesSlug) {
-      const seriesAuthors = getSeriesAuthors(effectiveSeriesSlug);
-      if (seriesAuthors) {
-        authors = seriesAuthors;
-      }
-    }
-    if (authors.length === 0) {
-      const defaultAuthors = siteConfig.posts?.authors?.default;
-      if (defaultAuthors && defaultAuthors.length > 0) {
-        authors = defaultAuthors;
-      }
-    }
-  }
-
-  const excerpt = data.excerpt || generateExcerpt(contentWithoutH1);
-  const readingMinutes = calculateReadingMinutes(contentWithoutH1);
-  const wordCount = calculateWordCount(contentWithoutH1);
-
-  let date = data.date;
-  if (!date && dateFromFileName) date = dateFromFileName;
-  if (!date) date = fs.statSync(fullPath).mtime.toISOString().split('T')[0];
-
-  const headings = getHeadings(content);
-
-  let coverImage = data.coverImage;
-  if (coverImage && !coverImage.startsWith('http') && !coverImage.startsWith('/') && !coverImage.startsWith('text:')) {
-    const cleanPath = coverImage.replace(/^\.\//, '');
-    coverImage = `/${imageBaseSlug}/${cleanPath}`;
-  }
-
-  return {
-    slug,
-    title: data.title,
-    subtitle: data.subtitle,
-    date,
-    excerpt,
-    category: data.category,
-    tags: data.tags,
-    authors,
-    layout: data.layout,
-    series: effectiveSeriesSlug,
-    seriesTitle: effectiveSeriesSlug ? getSeriesTitle(effectiveSeriesSlug) : undefined,
-    coverImage,
-    sort: data.sort,
-    posts: data.posts,
-    featured: data.featured,
-    pinned: data.pinned,
-    draft: data.draft,
-    latex: data.latex,
-    toc: data.toc,
-    commentable: data.commentable,
-    type: data.type,
-    items: data.items as CollectionItem[] | undefined,
-    externalLinks: data.externalLinks,
-    redirectFrom: data.redirectFrom,
-    readingMinutes,
-    wordCount,
-    content: contentWithoutH1,
-    headings,
-    imageBaseSlug,
-    sourceFormat: 'markdown',
-  };
-}
-
-export function parseMarkdownFileForTests(
-  fullPath: string,
-  slug: string,
-  dateFromFileName?: string,
-  seriesName?: string,
-): PostData {
-  return parseMarkdownFile(fullPath, slug, dateFromFileName, seriesName);
-}
-
-function parseRstFile(
-  fullPath: string,
-  slug: string,
-  dateFromFileName?: string,
-  seriesName?: string,
-  preRendered?: RenderedRstDocument,
-): PostData {
-  try {
-    const imageBaseSlug = getRstImageBaseSlug(fullPath, slug);
-    const fileContents = readUtf8File(fullPath);
-
-    let parsedTitle: string;
-    let parsedBody: string;
-    let parsedText: string | undefined;
-    let parsedHeadings: Heading[];
-    let parsedExcerpt: string;
-    let parsedReadingMinutes: number;
-    let parsedWordCount: number;
-    let parsedHtml: string | undefined;
-    let data: ReturnType<typeof parseRstDocument>['metadata'];
-    try {
-      if (preRendered) {
-        const rendered = preRendered;
-        parsedTitle = rendered.title;
-        parsedBody = rendered.text;
-        parsedText = rendered.text;
-        parsedHeadings = rendered.headings;
-        parsedExcerpt = rendered.excerpt;
-        parsedReadingMinutes = rendered.readingMinutes;
-        parsedWordCount = rendered.wordCount;
-        parsedHtml = rendered.html;
-        data = rendered.metadata;
-      } else if (shouldUsePythonRstRenderer() && pythonRstRendererAvailable !== false) {
-        const rendered = renderRstFile(fullPath, imageBaseSlug);
-        pythonRstRendererAvailable = true;
-        parsedTitle = rendered.title;
-        parsedBody = rendered.text;
-        parsedText = rendered.text;
-        parsedHeadings = rendered.headings;
-        parsedExcerpt = rendered.excerpt;
-        parsedReadingMinutes = rendered.readingMinutes;
-        parsedWordCount = rendered.wordCount;
-        parsedHtml = rendered.html;
-        data = rendered.metadata;
-      } else {
-        throw new Error('__RST_FALLBACK__');
-      }
-    } catch (error) {
-      if (!isPythonRuntimeUnavailable(error)) {
-        throw error;
-      }
-      if (pythonRstRendererAvailable !== false) {
-        pythonRstRendererAvailable = false;
-      }
-      const parsed = parseRstDocument(fileContents);
-      parsedTitle = parsed.title;
-      parsedBody = parsed.body;
-      parsedHeadings = parsed.headings;
-      parsedExcerpt = parsed.excerpt;
-      parsedReadingMinutes = parsed.readingMinutes;
-      parsedWordCount = parsed.wordCount;
-      data = parsed.metadata;
-    }
-
-    const effectiveSeriesSlug = data.series || seriesName;
-    let authors: string[] = [];
-    if (data.authors && data.authors.length > 0) {
-      authors = data.authors;
-    } else if (data.author) {
-      authors = [data.author];
-    } else {
-      if (effectiveSeriesSlug) {
-        const seriesAuthors = getSeriesAuthors(effectiveSeriesSlug);
-        if (seriesAuthors) authors = seriesAuthors;
-      }
-      if (authors.length === 0) {
-        const defaultAuthors = siteConfig.posts?.authors?.default;
-        if (defaultAuthors && defaultAuthors.length > 0) {
-          authors = defaultAuthors;
-        }
-      }
-    }
-
-    let date = data.date;
-    if (!date && dateFromFileName) date = dateFromFileName;
-    if (!date) date = fs.statSync(fullPath).mtime.toISOString().split('T')[0];
-
-    let coverImage = data.coverImage;
-    if (coverImage && !coverImage.startsWith('http') && !coverImage.startsWith('/') && !coverImage.startsWith('text:')) {
-      const cleanPath = coverImage.replace(/^\.\//, '');
-      coverImage = `/${imageBaseSlug}/${cleanPath}`;
-    }
-    const toctreePosts = isSeriesIndexRst(fullPath, slug, seriesName)
-      ? extractRstToctreePosts(fileContents)
-      : [];
-    const seriesPosts = data.posts && data.posts.length > 0
-      ? data.posts
-      : ((data.sort === undefined || data.sort === 'manual') && toctreePosts.length > 0 ? toctreePosts : undefined);
-    const sort = data.sort ?? (seriesPosts ? 'manual' : 'date-desc');
-
-    return {
-      slug,
-      title: parsedTitle,
-      subtitle: data.subtitle,
-      date,
-      excerpt: data.excerpt || parsedExcerpt,
-      category: data.category ?? 'Uncategorized',
-      tags: data.tags ?? [],
-      authors,
-      layout: data.layout ?? 'post',
-      series: effectiveSeriesSlug,
-      seriesTitle: effectiveSeriesSlug ? getSeriesTitle(effectiveSeriesSlug) : undefined,
-      coverImage,
-      sort,
-      posts: seriesPosts,
-      type: data.type,
-      featured: data.featured ?? false,
-      pinned: data.pinned ?? false,
-      draft: data.draft ?? false,
-      latex: data.latex ?? false,
-      toc: data.toc ?? true,
-      commentable: data.commentable,
-      redirectFrom: data.redirectFrom ?? [],
-      readingMinutes: parsedReadingMinutes,
-      wordCount: parsedWordCount,
-      content: parsedBody,
-      renderedHtml: parsedHtml,
-      plainText: parsedText,
-      headings: parsedHeadings,
-      imageBaseSlug,
-      sourceFormat: 'rst',
-    };
-  } catch (error) {
-    if (error instanceof RstParseError) {
-      throw new RstParseError(`${error.message} (${fullPath})`);
-    }
-    throw error;
-  }
-}
-
-export function parseRstFileForTests(
-  fullPath: string,
-  slug: string,
-  dateFromFileName?: string,
-  seriesName?: string,
-  preRendered?: RenderedRstDocument,
-): PostData {
-  return parseRstFile(fullPath, slug, dateFromFileName, seriesName, preRendered);
-}
-
-export function resetPythonRstRendererAvailabilityForTests(value: boolean | null = null): void {
-  pythonRstRendererAvailable = value;
-}
-
-export function getPythonRstRendererAvailabilityForTests(): boolean | null {
-  return pythonRstRendererAvailable;
-}
-
 export function getAllPosts(): PostData[] {
   const cacheKey = getCacheEnvKey();
   const cached = postsCache.get(cacheKey);
   if (cached) return cached;
 
   const allPostsData: PostData[] = [];
-  const pendingRstPosts: PendingRstPostEntry[] = [];
+  const pendingRstPosts: RstPostEntry[] = [];
 
   // Helper to process a directory
   const processDirectory = (dir: string, isSeriesDir: boolean = false) => {
@@ -753,39 +135,7 @@ export function getAllPosts(): PostData[] {
   processDirectory(contentDirectory);
   processDirectory(seriesDirectory, true);
 
-  if (pendingRstPosts.length > 0) {
-    let batchRenderedByFile: Map<string, RenderedRstDocument> | null = null;
-
-    if (shouldUsePythonRstRenderer() && pythonRstRendererAvailable !== false) {
-      try {
-        batchRenderedByFile = renderRstFilesBatch(
-          pendingRstPosts.map(entry => ({
-            file: entry.fullPath,
-            imageBaseSlug: getRstImageBaseSlug(entry.fullPath, entry.slug),
-          }))
-        );
-        pythonRstRendererAvailable = true;
-      } catch (error) {
-        if (isPythonRuntimeUnavailable(error)) {
-          pythonRstRendererAvailable = false;
-        } else {
-          throw error;
-        }
-      }
-    }
-
-    pendingRstPosts.forEach(entry => {
-      allPostsData.push(
-        parseRstFile(
-          entry.fullPath,
-          entry.slug,
-          entry.dateFromFileName,
-          entry.seriesSlug,
-          batchRenderedByFile?.get(entry.fullPath),
-        )
-      );
-    });
-  }
+  allPostsData.push(...parseRstPostEntries(pendingRstPosts));
 
   const result = allPostsData
     .filter(post => {
